requirements.txt
google-cloud-container
google-cloud-resource-manager
kubernetes
overrides~=6.1.0
rfc3987~=1.3.8
validators~=0.20.0
